requirements.txt
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
editor==1.6.6
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
inquirer==3.4.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
runs==1.2.2
six==1.17.0
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
xmod==1.8.1
